1. Class name adjusted to idValidator 2. Test class created to test method.
